refactor: adjust Campaign Donations block to the new implemented pattern
