+++ src/Campaigns/Blocks/CampaignDonations/app.tsx
@@ -1,0 +1,1 @@
+import './styles.scss';
