--- src/Campaigns/Blocks/CampaignDonationsBlock/style.css
@@ -1,190 +1,0 @@
-.givewp-campaign-donations-block {
-    padding: 1.5rem 0;
-}
-
-.givewp-campaign-donations-block * {
-    font-family: 'Inter', sans-serif;
-}
-
-.givewp-campaign-donations-block__header {
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 0.5rem;
-}
-
-.givewp-campaign-donations-block__title {
-    color: var(--givewp-neutral-900);
-    font-size: 1.125rem;
-    font-weight: 600;
-    line-height: 1.56;
-    margin: 0;
-}
-
-:is(.givewp-campaign-donations-block__donate-button, .givewp-campaign-donations-block__empty-button) button.givewp-donation-form-modal__open {
-    background: none;
-    border-radius: 0.5rem;
-    border: 1px solid var(--givewp-primary-color);
-    color: var(--givewp-primary-color) !important;
-    font-size: 0.875rem;
-    font-weight: 600;
-    line-height: 1.43;
-    padding: 0.25rem 1rem !important;
-}
-
-:is(.givewp-campaign-donations-block__donate-button, .givewp-campaign-donations-block__empty-button) button.givewp-donation-form-modal__open:hover {
-    background: var(--givewp-primary-color);
-    color: var(--givewp-shades-white) !important;
-}
-
-.givewp-campaign-donations-block__donations {
-    display: grid;
-    gap: 0.5rem;
-    margin: 0;
-    padding: 0;
-}
-
-.givewp-campaign-donations-block__donation {
-    align-items: center;
-    background-color: var(--givewp-shades-white);
-    border-radius: 0.5rem;
-    border: 1px solid var(--givewp-neutral-50);
-    display: flex;
-    gap: 0.75rem;
-    padding: 1rem;
-}
-
-.givewp-campaign-donations-block__donation-icon {
-    display: flex;
-    align-items: center;
-}
-
-.givewp-campaign-donations-block__donation-icon img {
-    border-radius: 100%;
-    height: 2.5rem;
-    object-fit: cover;
-    width: 2.5rem;
-}
-
-.givewp-campaign-donations-block__donation-info {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    row-gap: 0.25rem;
-}
-
-.givewp-campaign-donations-block__donation-description {
-    color: var(--givewp-neutral-500);
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 1.5;
-    margin: 0;
-}
-
-.givewp-campaign-donations-block__donation-description strong {
-    color: var(--givewp-neutral-700);
-    font-weight: 600;
-}
-
-.givewp-campaign-donations-block__donation-date {
-    align-items: center;
-    color: var(--givewp-neutral-400);
-    display: flex;
-    font-size: 0.875rem;
-    font-weight: 500;
-    line-height: 1.43;
-}
-
-.givewp-campaign-donations-block__donation-ribbon {
-    align-items: center;
-    border-radius: 100%;
-    color: #1F2937;
-    display: flex;
-    height: 1.25rem;
-    justify-content: center;
-    margin-left: auto;
-    width: 1.25rem;
-}
-
-.givewp-campaign-donations-block__donation-ribbon[data-position="1"] {
-    background-color: #ffd700;
-}
-
-.givewp-campaign-donations-block__donation-ribbon[data-position="2"] {
-    background-color: #c0c0c0;
-}
-
-.givewp-campaign-donations-block__donation-ribbon[data-position="3"] {
-    background-color: #cd7f32;
-    color: #fffaf2;
-}
-
-.givewp-campaign-donations-block__donation-amount {
-    color: var(--givewp-neutral-700);
-    font-size: 1.125rem;
-    font-weight: 600;
-    line-height: 1.56;
-    margin-left: auto;
-}
-
-.givewp-campaign-donations-block__footer {
-    display: flex;
-    justify-content: center;
-    margin-top: 0.5rem;
-}
-
-.givewp-campaign-donations-block__load-more-button {
-    background: none;
-    border-radius: 0.5rem;
-    border: 1px solid var(--givewp-primary-color);
-    color: var(--givewp-primary-color);
-    font-size: 0.875rem;
-    font-weight: 600;
-    line-height: 1.43;
-    padding: 0.25rem 1rem;
-}
-
-.givewp-campaign-donations-block__load-more-button:hover {
-    background: var(--givewp-primary-color);
-    color: var(--givewp-shades-white);
-}
-
-.givewp-campaign-donations-block__empty-state {
-    align-items: center;
-    background-color: var(--givewp-shades-white);
-    border-radius: 0.5rem;
-    border: 1px solid var(--givewp-neutral-50);
-    display: flex;
-    flex-direction: column;
-    padding: 1.5rem;
-}
-
-.givewp-campaign-donations-block__empty-title {
-    color: var(--givewp-neutral-700);
-    font-size: 1rem;
-    font-weight: 500;
-    line-height: 1.5;
-    margin: 0;
-}
-
-.givewp-campaign-donations-block__empty-description {
-    color: var(--givewp-neutral-700);
-    font-size: 0.875rem;
-    line-height: 1.43;
-    margin: 0.25rem 0 0;
-}
-
-.givewp-campaign-donations-block__empty-icon {
-    color: var(--givewp-secondary-color);
-    margin-bottom: 0.875rem;
-    order: -1;
-}
-
-.givewp-campaign-donations-block__empty-button {
-    margin-top: 0.875rem;
-}
-
-.givewp-campaign-donations-block__empty-button button.givewp-donation-form-modal__open {
-    border-radius: 0.25rem;
-    padding: 0.5rem 1rem !important;
-}
